Add parameter to discard checksum control
Fixes #20

Added extra parameter, discard_checksum, defaulting to False. If set,
simply discard the checksum control.

Fixes #25

Search for spdxdocument, instead of building up the filename which
would imply a requirement on the name of the SPDX files which is not
nice, If file not found in the SPDX dirs, raise exeption.